休憩時間のバリデーション、入力時間を分単位に修正

+++ db/migrate/20211102025233_create_records.rb
@@ -4,7 +4,7 @@ class CreateRecords < ActiveRecord::Migration[6.0]
       t.date :day_time
       t.time :begin_time
       t.time :finish_time
-      t.float :break_time
+      t.integer :break_time
       t.references :user, foreign_key: true, null: false
       t.timestamps
     end
